Moved receipt package to the top payments package
