+++ WooCommerce/src/main/res/navigation/nav_graph_orders.xml
@@ -488,7 +488,7 @@
     </fragment>
     <fragment
         android:id="@+id/receiptPreviewFragment"
-        android:name="com.woocommerce.android.ui.payments.cardreader.receipt.ReceiptPreviewFragment"
+        android:name="com.woocommerce.android.ui.payments.receipt.ReceiptPreviewFragment"
         android:label="ReceiptPreviewFragment"
         tools:layout="@layout/fragment_receipt_preview">
         <argument
